migrate to db sync 13.0.0

+++ modules/api/src/main/scala/io/ergolabs/cardano/explorer/api/v1/services/NetworkParamsService.scala
@@ -34,7 +34,7 @@ object NetworkParamsService {
         epochParams     <- repos.network.getLastEpochParams
         costModel       <- repos.network.getCostModel(epochParams.costModelId)
         parsedCm        <- parser.parse(costModel).toRaise
-        transformed     <- parsedCm.as[Map[String, Map[String, Int]]].toRaise
+        transformed     <- parsedCm.as[Map[String, Map[String, Long]]].toRaise
         cmCorrectFormat = transformed.map { case (k, v) => "PlutusScriptV1" -> v }
       } yield 
           EnvParams(
